refactor: extract generation result panels
