+++ frontend/src/App.jsx
@@ -11,6 +11,9 @@ import AppHeader, { SignInDialog } from "./components/layout/AppHeader";
 import ContextInputsPanel from "./components/context/ContextInputsPanel";
 import ExportPanel from "./components/export/ExportPanel";
 import GeneratedTestCasesView from "./components/generation/GeneratedTestCasesView";
+import RequirementAnalysisPanel from "./components/generation/RequirementAnalysisPanel";
+import ScenarioCoveragePanel from "./components/generation/ScenarioCoveragePanel";
+import TraceabilityMatrixPanel from "./components/generation/TraceabilityMatrixPanel";
 import BillingBanner from "./components/layout/BillingBanner";
 import RequirementReviewWorkbench from "./components/requirements/RequirementReviewWorkbench";
 import SettingsDialog from "./components/settings/SettingsDialog";
@@ -38,7 +41,6 @@ import { API_BASE, createRequestId, downloadResponseBlob, parseApiError } from "
 import {
 	buildAzureDevOpsConnectionForm,
 	buildJiraConnectionForm,
-	getRequirementContextPath,
 	getRequirementReviewStatus,
 	getTestCaseLinkedRequirementIds,
 	isAzureDevOpsLinkedRequirement,
@@ -2806,264 +2808,37 @@ export default function App() {
 										)
 									)}
 
-									{activeGenerateResultTab === "coverage" && (coveragePlan.length > 0 ? (
-							<div className="result-section">
-										<details className="collapsible-panel" open>
-									<summary className="collapsible-panel-summary">
-										<span className="collapsible-panel-copy">
-											<span className="collapsible-panel-title">Scenario Coverage Plan</span>
-											<span className="collapsible-panel-description">
-												Planned scenario intent per requirement, available on demand instead of taking over the page.
-											</span>
-										</span>
-										<span className="collapsible-panel-meta">
-											<span className="analysis-summary-pill">{coveragePlan.length} requirements</span>
-											<span className="analysis-summary-pill">Scenarios {coveredScenarioTotal}/{plannedScenarioTotal}</span>
-											<span className="analysis-summary-pill">Must-have {mustHaveCoveredScenarioTotal}/{mustHaveScenarioTotal}</span>
-											{missingScenarioCount > 0 && (
-												<span className="analysis-summary-pill collapsible-pill-alert">Missing {missingScenarioCount}</span>
-											)}
-											<span className="collapsible-panel-icon" aria-hidden="true">⏄</span>
-										</span>
-									</summary>
-									<div className="collapsible-panel-body">
-										<div className="coverage-plan-list">
-											{coveragePlan.map((plan) => {
-												const summary = getRequirementScenarioSummary(plan.requirement_id);
-												const missingScenarioTypes = new Set(summary?.missing_scenario_types || []);
-												return (
-													<div key={plan.requirement_id} className="coverage-plan-card">
-														<div className="coverage-plan-header">
-															<div>
-																<div className="coverage-plan-id">{plan.requirement_id}</div>
-																<div className="coverage-plan-text">{plan.requirement_text}</div>
-															</div>
-															{summary && (
-																<span className="coverage-plan-summary">
-																	{summary.covered_scenarios}/{summary.planned_scenarios} planned scenarios covered
-																</span>
-															)}
-														</div>
-														<div className="coverage-chip-row">
-															{plan.scenarios?.map((scenario) => {
-																const isMissing = missingScenarioTypes.has(scenario.scenario_type);
-																return (
-																	<span
-																		key={scenario.id}
-																		className={`coverage-chip ${scenario.must_have ? "required" : "recommended"} ${isMissing ? "missing" : "covered"}`}
-																		title={scenario.objective}
-																	>
-																		{scenario.scenario_type}
-																	</span>
-																);
-															})}
-														</div>
-													</div>
-												);
-											})}
-										</div>
-									</div>
-								</details>
-							</div>
-									) : (
-										<div className="generate-result-empty">
-											<h3>Scenario Coverage Plan</h3>
-											<p>No scenario coverage plan is available for this run.</p>
-										</div>
-									))}
+									{activeGenerateResultTab === "coverage" && (
+										<ScenarioCoveragePanel
+											coveragePlan={coveragePlan}
+											coveredScenarioTotal={coveredScenarioTotal}
+											plannedScenarioTotal={plannedScenarioTotal}
+											mustHaveCoveredScenarioTotal={mustHaveCoveredScenarioTotal}
+											mustHaveScenarioTotal={mustHaveScenarioTotal}
+											missingScenarioCount={missingScenarioCount}
+											getRequirementScenarioSummary={getRequirementScenarioSummary}
+										/>
+									)}
 
-									{activeGenerateResultTab === "analysis" && (requirementAnalysis.length > 0 ? (
-							<div className="result-section">
-										<details className="collapsible-panel" open>
-									<summary className="collapsible-panel-summary">
-										<span className="collapsible-panel-copy">
-											<span className="collapsible-panel-title">Requirement Analysis</span>
-											<span className="collapsible-panel-description">
-												Rules, constraints, permissions, transitions, and risks extracted before scenario planning.
-											</span>
-										</span>
-										<span className="collapsible-panel-meta">
-											<span className="analysis-summary-pill">{requirementAnalysis.length} requirements</span>
-											{coverageMetrics && (
-												<>
-													<span className="analysis-summary-pill">Rules {coverageMetrics.business_rules_covered || 0}/{coverageMetrics.business_rules_total || 0}</span>
-													<span className="analysis-summary-pill">Constraints {coverageMetrics.field_constraints_covered || 0}/{coverageMetrics.field_constraints_total || 0}</span>
-												</>
-											)}
-											{requirementAnalysisGapCount > 0 && (
-												<span className="analysis-summary-pill collapsible-pill-alert">Gaps {requirementAnalysisGapCount}</span>
-											)}
-											<span className="collapsible-panel-icon" aria-hidden="true">⏄</span>
-										</span>
-									</summary>
-									<div className="collapsible-panel-body">
-										{coverageMetrics && (
-											<div className="analysis-overview-row">
-												<span className="analysis-summary-pill">Rules {coverageMetrics.business_rules_covered || 0}/{coverageMetrics.business_rules_total || 0}</span>
-												<span className="analysis-summary-pill">Constraints {coverageMetrics.field_constraints_covered || 0}/{coverageMetrics.field_constraints_total || 0}</span>
-												<span className="analysis-summary-pill">Permissions {coverageMetrics.role_permissions_covered || 0}/{coverageMetrics.role_permissions_total || 0}</span>
-												<span className="analysis-summary-pill">Transitions {coverageMetrics.state_transitions_covered || 0}/{coverageMetrics.state_transitions_total || 0}</span>
-												<span className="analysis-summary-pill">Risks {coverageMetrics.risk_signals_covered || 0}/{coverageMetrics.risk_signals_total || 0}</span>
-											</div>
-										)}
-										<div className="analysis-card-list">
-											{requirementAnalysis.map((analysis) => {
-												const summary = getRequirementAnalysisSummary(analysis.requirement_id);
-												const gaps = getRequirementAnalysisGaps(analysis.requirement_id);
-												const hasGaps = Object.values(gaps).some((items) => items.length > 0);
-												return (
-													<div key={analysis.requirement_id} className="analysis-card">
-														<div className="analysis-card-header">
-															<div>
-																<div className="coverage-plan-id">{analysis.requirement_id}</div>
-																<div className="coverage-plan-text">{analysis.requirement_text}</div>
-															</div>
-															{summary && (
-																<span className="coverage-plan-summary">
-																	{summary.business_rules_covered}/{summary.business_rules_total} rules • {summary.field_constraints_covered}/{summary.field_constraints_total} constraints
-																</span>
-															)}
-														</div>
-														<div className="analysis-summary-row">
-															<span className="analysis-summary-pill">Rules {analysis.business_rules?.length || 0}</span>
-															<span className="analysis-summary-pill">Constraints {analysis.field_constraints?.length || 0}</span>
-															<span className="analysis-summary-pill">Permissions {analysis.role_permissions?.length || 0}</span>
-															<span className="analysis-summary-pill">Transitions {analysis.state_transitions?.length || 0}</span>
-															<span className="analysis-summary-pill">Risks {analysis.risk_signals?.length || 0}</span>
-														</div>
-														{analysis.suggested_scenarios?.length > 0 && (
-															<div className="analysis-chip-row">
-																{analysis.suggested_scenarios.map((scenario) => (
-																	<span key={`${analysis.requirement_id}-${scenario}`} className="analysis-chip">
-																		{scenario}
-																	</span>
-																))}
-															</div>
-														)}
-														<div className="analysis-detail-grid">
-															<div className="analysis-detail-block">
-																<h4>Business rules</h4>
-																<ul className="analysis-detail-list">
-																	{(analysis.business_rules || []).slice(0, 2).map((rule) => (
-																		<li key={rule.id}>{rule.title}</li>
-																	))}
-																</ul>
-															</div>
-															<div className="analysis-detail-block">
-																<h4>Constraints</h4>
-																<ul className="analysis-detail-list">
-																	{(analysis.field_constraints || []).slice(0, 2).map((constraint) => (
-																		<li key={constraint.id}>{constraint.field_name}: {constraint.description}</li>
-																	))}
-																</ul>
-															</div>
-															<div className="analysis-detail-block">
-																<h4>Permissions</h4>
-																<ul className="analysis-detail-list">
-																	{(analysis.role_permissions || []).slice(0, 2).map((permission) => (
-																		<li key={permission.id}>{permission.role}: {permission.action}</li>
-																	))}
-																</ul>
-															</div>
-															<div className="analysis-detail-block">
-																<h4>Transitions</h4>
-																<ul className="analysis-detail-list">
-																	{(analysis.state_transitions || []).slice(0, 2).map((transition) => (
-																		<li key={transition.id}>{transition.from_state} → {transition.to_state}</li>
-																	))}
-																</ul>
-															</div>
-															<div className="analysis-detail-block">
-																<h4>Risks</h4>
-																<ul className="analysis-detail-list">
-																	{(analysis.risk_signals || []).slice(0, 2).map((risk) => (
-																		<li key={risk.id}>{risk.severity}: {risk.title}</li>
-																	))}
-																</ul>
-															</div>
-														</div>
-														{hasGaps && (
-															<div className="analysis-gap-block">
-																<strong>Coverage gaps</strong>
-																<ul className="analysis-gap-list">
-																	{gaps.highRisks.slice(0, 2).map((item) => <li key={item}>{item}</li>)}
-																	{gaps.rules.slice(0, 2).map((item) => <li key={item}>{item}</li>)}
-																	{gaps.constraints.slice(0, 2).map((item) => <li key={item}>{item}</li>)}
-																	{gaps.permissions.slice(0, 2).map((item) => <li key={item}>{item}</li>)}
-																	{gaps.transitions.slice(0, 2).map((item) => <li key={item}>{item}</li>)}
-																</ul>
-															</div>
-														)}
-													</div>
-												);
-											})}
-										</div>
-									</div>
-								</details>
-							</div>
-									) : (
-										<div className="generate-result-empty">
-											<h3>Requirement Analysis</h3>
-											<p>No requirement analysis is available for this run.</p>
-										</div>
-									))}
+									{activeGenerateResultTab === "analysis" && (
+										<RequirementAnalysisPanel
+											requirementAnalysis={requirementAnalysis}
+											coverageMetrics={coverageMetrics}
+											requirementAnalysisGapCount={requirementAnalysisGapCount}
+											getRequirementAnalysisSummary={getRequirementAnalysisSummary}
+											getRequirementAnalysisGaps={getRequirementAnalysisGaps}
+										/>
+									)}
 
-									{activeGenerateResultTab === "traceability" && (approvedRequirements.length > 0 ? (
-							<div className="result-section">
-								<h3>Traceability Matrix</h3>
-								<div className="workflow-diagnostics-pills">
-									<span className="workflow-diagnostics-pill">Approved requirements covered {tracedRequirementCount}/{approvedRequirements.length}</span>
-									<span className="workflow-diagnostics-pill">Cases with traceability {coverageMetrics?.cases_with_traceability ?? 0}/{testCases.length}</span>
-									<span className="workflow-diagnostics-pill">Scenario coverage {coverageMetrics?.covered_planned_scenarios ?? 0}/{coverageMetrics?.planned_scenarios_total ?? 0}</span>
-								</div>
-								<div className="traceability-table-wrapper">
-									<table className="traceability-table">
-										<thead>
-											<tr>
-												<th>Requirement</th>
-												<th>Story / source path</th>
-												<th>Linked test cases</th>
-												<th>Scenario coverage</th>
-												<th>Status</th>
-											</tr>
-										</thead>
-										<tbody>
-											{requirementTraceabilityRows.map(({ requirement, linkedTestCases, scenarioSummary, linkedScenarioTypes }) => {
-												const isCovered = linkedTestCases.length > 0;
-												return (
-													<tr key={requirement.id} className={isCovered ? "covered" : "missing"}>
-														<td>
-															<strong>{requirement.id}</strong>
-															<span>{requirement.text}</span>
-														</td>
-														<td>{getRequirementContextPath(requirement)}</td>
-														<td>
-															{linkedTestCases.length ? linkedTestCases.map((testCase) => (
-																<span key={testCase.id} className="tag traceability-case-tag">{testCase.id}</span>
-															)) : <span className="traceability-missing-text">No linked tests</span>}
-														</td>
-														<td>
-															{scenarioSummary ? `${scenarioSummary.covered_scenarios}/${scenarioSummary.planned_scenarios}` : "—"}
-															{linkedScenarioTypes.length > 0 && (
-																<div className="traceability-scenario-tags">
-																	{linkedScenarioTypes.slice(0, 4).map((scenario) => <span key={`${requirement.id}-${scenario}`}>{scenario}</span>)}
-																</div>
-															)}
-														</td>
-														<td><span className={`traceability-status ${isCovered ? "covered" : "missing"}`}>{isCovered ? "Covered" : "Gap"}</span></td>
-													</tr>
-												);
-											})}
-										</tbody>
-									</table>
-								</div>
-							</div>
-									) : (
-										<div className="generate-result-empty">
-											<h3>Traceability Matrix</h3>
-											<p>No approved requirements are available to trace for this run.</p>
-										</div>
-									))}
+									{activeGenerateResultTab === "traceability" && (
+										<TraceabilityMatrixPanel
+											approvedRequirements={approvedRequirements}
+											requirementTraceabilityRows={requirementTraceabilityRows}
+											tracedRequirementCount={tracedRequirementCount}
+											coverageMetrics={coverageMetrics}
+											testCases={testCases}
+										/>
+									)}
 
 									{activeGenerateResultTab === "test-cases" && (
 										<GeneratedTestCasesView
